+++ R/rticulate.R
@@ -1,7 +1,7 @@
-#' rticulate: A small package for Ultrasound Tongue Imaging in R.
+#' rticulate: Ultrasound Tongue Imaging in R.
 #'
-#' This is a package providing a function for processing AAA splines
-#' files and plot tongue contours.
+#' This package provides function for processing Articulate Assistant Advanced
+#' (AAA) export files and plot tongue contour data from any system.
 #'
 #' @docType package
 #' @name rticulate