+++ NEWS.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## Unreleased
+### Added
+- argument in `read_aaa` to chose coordinate system
+
 ### Changed
 - updates to `NEWS.md`