+++ NEWS.md
@@ -12,6 +12,7 @@
 
 ### Fixed
 - name of package in documentation
+- name of vignette
 
 ## [1.0.1] - 2017-06-10
 ### Added